requirements.txt
streamlit
playwright
openai
chromadb
python-dotenv
numpy
requests
beautifulsoup4